Fix imports after renames
