+++ tests/unit/states/test_boto_elasticsearch_domain.py
@@ -25,7 +25,7 @@ import salt.loader
 from salt.ext.six.moves import range  # pylint: disable=import-error,redefined-builtin
 
 # pylint: disable=import-error,no-name-in-module,unused-import
-from unit.modules.boto_elasticsearch_domain_test import BotoElasticsearchDomainTestCaseMixin
+from unit.modules.test_boto_elasticsearch_domain import BotoElasticsearchDomainTestCaseMixin
 
 # Import 3rd-party libs
 try:
