+++ tests/unit/states/test_boto_s3_bucket.py
@@ -26,7 +26,7 @@ import salt.loader
 from salt.ext.six.moves import range  # pylint: disable=import-error,redefined-builtin
 
 # pylint: disable=import-error,no-name-in-module,unused-import
-from unit.modules.boto_s3_bucket_test import BotoS3BucketTestCaseMixin
+from unit.modules.test_boto_s3_bucket import BotoS3BucketTestCaseMixin
 
 # Import 3rd-party libs
 try:
