+++ tests/unit/transport/test_zeromq.py
@@ -33,8 +33,8 @@ ensure_in_syspath('../')
 import integration
 
 # Import Salt libs
-from unit.transport.req_test import ReqChannelMixin
-from unit.transport.pub_test import PubChannelMixin
+from unit.transport.test_req import ReqChannelMixin
+from unit.transport.test_pub import PubChannelMixin
 
 ON_SUSE = False
 if 'SuSE' in platform.dist():
